Issue 4513 get rid of util class
