+++ api/src/main/java/com/epam/pipeline/manager/cluster/performancemonitoring/monitor/CpuIdleRunMonitor.java
@@ -24,7 +24,6 @@ import com.epam.pipeline.entity.monitoring.IdleMonitoringConfig;
 import com.epam.pipeline.entity.monitoring.IdleMonitoringType;
 import com.epam.pipeline.entity.pipeline.PipelineRun;
 import com.epam.pipeline.manager.cluster.performancemonitoring.monitor.common.InstanceTypeCache;
-import com.epam.pipeline.manager.cluster.performancemonitoring.monitor.common.RunMonitorUtils;
 import com.epam.pipeline.manager.notification.NotificationManager;
 import com.epam.pipeline.manager.pipeline.PipelineRunDockerOperationManager;
 import com.epam.pipeline.manager.pipeline.PipelineRunManager;
@@ -69,7 +68,7 @@ public class CpuIdleRunMonitor extends AbstractIdleRunMonitor {
 
     @Override
     public void monitor(final List<PipelineRun> runs) {
-        final Map<String, PipelineRun> running = RunMonitorUtils.groupedByNode(runs, messageHelper);
+        final Map<String, PipelineRun> running = groupedByNode(runs);
         final IdleMonitoringConfig conf = findEnabledIdleConfig(getType());
         if (!isIdleConfigReadyForProcessing(conf, getType())) {
             return;
